+++ src/drivers/tcp.h
@@ -12,7 +12,7 @@
  * backs normal guest services such as tcpecho, sockeof, and ftpd.
  */
 
-void tcp_init(void);
+int  tcp_init(void);
 int  tcp_handle_ipv4_frame(const unsigned char* frame, unsigned int len);
 void tcp_socket_handle_close(fd_entry_t* ent);
 void tcp_socket_use_port(unsigned int port);